Splitter opp schedule og transaction då optimistic lock exception kastes når spring-transaksjonen skal committes
